scripts/requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9